requirements.txt
pandas==0.22.0
Pillow==5.1.0
requests==2.18.4
xlrd==1.1.0